LayoutTests/platform/ios-sim-deprecated/http/tests/security/xss-DENIED-htmlelelment-with-iframe-proto-expected.txt

CONSOLE MESSAGE: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
 Tests that making other frame window a prototype doesn't expose that window subframe
PASS: targetWindow.myinput should be 'undefined' and is.
